requirements.txt
alembic==1.11.1
asyncpg==0.28.0
fastapi==0.100.0
sqlmodel==0.0.8
uvicorn==0.22.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
# Load testing - locust is required only for testing and it can be removed from main API image
locust==2.5.1
httpx==0.23.0
boto3==1.35.84
botocore==1.35.84
flask-cors==4.0.1
flask_caching==2.3.0
flask==3.0.3
pytest
aiosqlite
snowflake-connector-python==3.12.1
google-auth==2.34.0
google-cloud-storage==2.18.2
google-cloud-bigquery==3.25.0
google-cloud-secret-manager==2.20.2
databricks-sql-connector==3.4.0
email_validator==2.2.0
pycryptodome==3.18.0
pkcs7==0.1.2
python-multipart==0.0.9
# jsonschema==4.21.1
grpcio==1.65.4
grpcio-tools==1.65.4
protobuf==5.27.3
pgvector==0.3.6
numpy==1.25.2
git+https://${TOKEN}@github.com/bh-ai/bh-cluster-utils.git@main#egg=bh-cluster-utils
git+https://${TOKEN}@github.com/bh-ai/bh-flow-service.git@main#egg=bh-flow-service
celery==5.4.0
oracledb==2.5.1
aiomysql==0.2.0
redis==5.2.0
